Update TLS Docs to represent new TLS enable command
Previously, enabling TLS required a user to create a service key in
order to find the hostname of the RMQ instance. As of V1.16, this is no
longer the case: users need only provide the command-line args '"tls":
true'.

This arg can also now be provided when the service is created, not just
on service update. This commit fixes the docs to better represent this
change.

[#164380997]